feat(editor): Refactoring time #36
