+++ src/app/models/strat-editor/editor-layer.spec.ts
@@ -1,0 +1,7 @@
+import { EditorLayer } from './editor-layer';
+
+describe('EditorLayer', () => {
+  it('should create an instance', () => {
+    expect(new EditorLayer()).toBeTruthy();
+  });
+});
