+++ src/app/models/strat-editor/utility.ts
